move searializer out to its own class
